pyproject.toml
[project]
name = "autonomous-skill-agent"
version = "0.1.0"
description = "Autonomous skill-native agent with progressive disclosure and multi-provider LLM routing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.8.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pyyaml>=6.0.1",
  "rapidfuzz>=3.9.0",
  "anthropic>=0.34.0",
  "google-genai>=0.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-mock>=3.14.0",
  "syrupy>=4.6.0",
  "ruff>=0.6.0",
  "pyright>=1.1.380",
]

[project.scripts]
agent = "agent.cli:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "N", "RUF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportMissingParameterType = false
reportUnknownArgumentType = false
reportMissingImports = false
exclude = [".venv", "runs", "demos", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
